pyproject.toml
[project]
name = "anco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "addict>=2.4.0",
    "loguru>=0.7.3",
    "openai>=1.107.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
